% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Wskazniki emisji z transportu drogowego EEA}
\format{
Obiekt zawiera
\describe{
\item{Category}{Kategoria pojazdu}
\item{Fuel}{Rodzaj paliwa}
\item{Segment}{Segment}
\item{Euro.Standard}{Standard emisji EURO}
\item{Technology}{Technologia}
\item{Pollutant}{Zanieczyszczenie}
\item{Mode}{Tryb}
\item{Alpha}{Parametr potrzebny od obliczen}
\item{Beta}{Parametr potrzebny od obliczen}
\item{Gamma}{Parametr potrzebny od obliczen}
\item{Delta}{Parametr potrzebny od obliczen}
\item{Epsilon}{Parametr potrzebny od obliczen}
\item{Zita}{Parametr potrzebny od obliczen}
\item{Hta}{Parametr potrzebny od obliczen}
\item{Reduction}{Redukcja}
\item{Bio}{Wskaznik BIO}
\item{Procent}{Procent}
}
}
\source{
MS Teams zajec
}
\usage{
wskazniki
}
\description{
Wskazniki emisji z transportu drogowego EEA
}
\examples{
\dontrun{
 wskazniki
 input
}
}
\keyword{datasets}
